Add node_subpage url redirect migration and cleanup
Delete config from jcc_tc_migration because it exists in the jcc_tc profile and we want
to be able to enable the module on v2 sites without getting all that config.

Delete the web/priate/scripts/updates/updates.php as it's a left over quicksilver script we don't use.

Update sc config.