Troubleshooting php version on Live.
The change from php 7.3 to 7.4 in pantheon.yml is not trigging the change on the live environment.
I've disabled the live tag on deploy to master so I can reset dev and test and then try to deploy
the php version change through the normal pantheon serial process.

The version updates on all other environemnts except Live, so I'm trying to elmiminate the unorthodox
hotfix style deploy to live as the cause.